fix/rest-client-routing: dynamic url for rest client
