+++ src/app/[locale]/rest-client/[...slug]/page.tsx
@@ -1,0 +1,5 @@
+'user client'
+
+export default function Page() {
+  return <p></p>
+}
